+++ app/controllers/legal_panel_for_government/rm6360/admin/reports_controller.rb
@@ -1,0 +1,9 @@
+module LegalPanelForGovernment
+  module RM6360
+    module Admin
+      class ReportsController < LegalPanelForGovernment::Admin::FrameworkController
+        include ::Admin::ReportsController
+      end
+    end
+  end
+end