+++ app/controllers/legal_services/rm6240/admin/dashboard_controller.rb
@@ -1,0 +1,8 @@
+module LegalServices
+  module RM6240
+    module Admin
+      class DashboardController < LegalServices::Admin::DashboardController
+      end
+    end
+  end
+end