+++ app/controllers/management_consultancy/rm6187/admin/uploads_controller.rb
@@ -2,7 +2,6 @@ module ManagementConsultancy
   module RM6187
     module Admin
       class UploadsController < ManagementConsultancy::Admin::FrameworkController
-        include SharedPagesConcern
         include ::Admin::UploadsController
 
         private