+++ spec/controllers/management_consultancy/rm6309/admin/sessions_controller_spec.rb
@@ -109,8 +109,8 @@ RSpec.describe ManagementConsultancy::RM6309::Admin::SessionsController do
       context 'and there is no challenge' do
         let(:challenge_name) { nil }
 
-        it 'redirects to management_consultancy_rm6309_admin_uploads_path' do
-          expect(response).to redirect_to management_consultancy_rm6309_admin_uploads_path
+        it 'redirects to management_consultancy_rm6309_admin_path' do
+          expect(response).to redirect_to management_consultancy_rm6309_admin_path
         end
       end
 
@@ -151,8 +151,8 @@ RSpec.describe ManagementConsultancy::RM6309::Admin::SessionsController do
         cookies.update(response.cookies)
       end
 
-      it 'redirects to management_consultancy_rm6309_admin_uploads_path' do
-        expect(response).to redirect_to management_consultancy_rm6309_admin_uploads_path
+      it 'redirects to management_consultancy_rm6309_admin_path' do
+        expect(response).to redirect_to management_consultancy_rm6309_admin_path
       end
     end
   end