+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Updated the admin sections for the legacy services to include the ability to generate reports of user activity ([PR 2034](https://github.com/Crown-Commercial-Service/crown-marketplace-legacy/pull/2034))
+
 ## [6.0.0] - 2025-08-07
 
 ### Added